cmd/snap/auto-import: stop importing system user assertions from initramfs mnts
These are erroneously being imported, we don't want assertions on the partitions
ubuntu-seed, ubuntu-boot, or ubuntu-data to be imported automatically, but the
current mount code sees these disks as already mounted since they are mounted in
the initramfs and as such considered for auto-import.

A more intelligent way to handle this would be to follow the bind mounts to a
device and see if those devices correspond to any of the partitions, but this
should be good enough as /var/lib/snapd/seed should never be a mount point on
Ubuntu Core 16 or 18.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>