cmd/s-b/initramfs-mounts: apply gadget defaults in install mode
Also don't load the seed again, load it once in the common code path for recover
and install modes so that we can reuse the gadget snap we loaded from the seed.

Also add unit tests for the install case and check that cloud-init was disabled
in recover mode too.

Finally make it explicit that we disallow cloud-init from running in these modes
with AllowCloudInit: false (this was already the case from default values, but
was not clear reading the code).

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>