New tests.backup tool
The new tool is used for backup and restore the test fs.

The meachnism is already being used by the prepare/restore for tests on
main suite, but now it is extended to be used in nested tests as well.

The idea of the tool is to backup the test directory revious to the test
execution and restore the backup after hte test is restored.