o/snapstate/autorefresh.go: eliminate race when launching autorefresh
When ensuring that an autorefresh is held via console-conf-start, we need to get
a state lock and set the refresh hold, but if another task in overlord already
started an auto-refresh, via autoRefresh.Ensure(), it will at some point release
the lock because it does network operations after checking whether the
auto-refresh should be held back or not, effectively meaning that the hold gets
ignored. To remedy this, we should check whether there is an effective hold
again right before actually queueing up the change to be run by the overlord.

Also add a unit test for this situation.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>