snapcraft.yaml: allow configuring the snapd snap build via dev options files
Presently, snapcraft does not allow specifying any environment variables or
anything that can be read or used from inside the snapcraft.yaml to customize
the build, so to accomplish this we use a gitignored top-level directory,
developer-snapd-snap-options which will be copied from the host's system into
the build environment when building a snap, to put files that act like env vars
to customize the build appropriately.

This commit also exposes the "testkeys" and "both-testkeys" options with the
former building a snapd snap that directly uses testkey root of trust.
The latter building a snapd snap with normal root of trust, but also including
the testkeys build snapd deb for CI purposes, as building both like this
minimizes the amount of time we spend building things to get both a normal snapd
snap that folks can download from the PR and use without compromising their
system's trust, and also be able to easily repack into a version that uses the
testkeys snapd for testing.

Relevant timings on my build machine for building the snapd snap with these
options turned on:

testkeys:       12m25.434s
both-testkeys:  15m3.530s + (< 1m for repacking)

Thus, if we just duplicated the whole snap build, it would take ~25 minutes,
whereas building both the testkeys deb and the normal one inside the snapcraft
build env saves us a whole 9-10 minutes. Note that is probably due to all the
overhead of building other things inside the snapd snap, as well as the time it
takes to stand up/setup the build environment by snapcraft.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>