client/client.go: addt'l verification of doOptions + unit tests for doOptions
This adds more verification to the doOptions passed to client.Do(), as well
as exports doOptions as DoOptions in export_test.go so we can write specific
unit tests around the verification and the retry implementation.

Implementing a unit test for the timeout behavior would be nice, but it's
unclear how to implement that right now without using "real" clock time,
which is susceptible to flakiness on slower hardware and is generally not
good to use in tests.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>