seed/seed20.go: check for devmode snaps, enable in Seed.Snap if allowed
For grade dangerous models, if a snap is in devmode confinement, then we should
specify that in the Seed.Snap we return from addSnap, which gets bubbled up to
seeding on first boot.

For extra paranoia, add additional checks around when we process snap infos, and
make sure we don't allow devmode snaps when the options.yaml didn't say so, and
also when the grade is not model dangerous.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>