snapshotstate: add clenup of abandonded snapshot imports
When a snapshot import is going on there is a chance that this
gets interrupted because e.g. a reboot is happening or snapd
crashes or restarts. This will leave incomplete import files
around.

This commit will clean those files on startup of snapd and
also adds flock based locks around the importing to make sure
the cleanup is reliable and not racy.