tests: add tests.cleanup pop sub-command
The "tests.cleanup pop" command pops the last command from the
cleanup stack and executes it. This functionality is useful for writing
tests where we want to use the "tests.cleanup restore" as a safeguard,
to ensure that, even in case of failure, the cleanup operation will be
performed, but want to perform the same cleanup faster in the case of
success.

For example:

    prepare: |
	tests.cleanup restore
    execute: |
	1) systemctl start helper.service
	2) tests.cleanup defer systemctl stop helper.service
	3) ... (some test activity that may fail)
	4) tests.cleanup pop
	5) ... (test code that should run without helper service)
    restore: |
	tests.cleanup restore

This allows us to avoid having to repeat the cleanup command twice and
diverging the cleanup logic into success and failure flows, one of which
would be usually not exercised.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>