many: have install return encryption keys for data and save, improve tests
Refactor install.Run() to return state information about the installed system.
In particular, return the encryption/reovery keys for ubuntu-data and
ubuntu-save partitions. With this, the saving of keys has been moved to
devicestate, what allowed for further improvements in unit tests, to verify that
the keys are properly saved to disk.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>