tests/lib/fakestore: wire up --actions-endpoint-contextual for running fakestore
This will be passed along to the handler for the v2 refresh action endpoint,
which then dictates that the handler shall operate like the real store and
return things that are actually in context of the client's context, meaning that
if the client says they are on revision 2, and the fakestore does not have a
revision of that snap above 2, then no refresh is provided, but if there is a
higher revision available, then that one is provided instead.

This commit is just the mechanical option passing, the next commits will flesh
out the implementation.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>