tests/lib/fakestore: add hard-coded nonce and session endpoints
This is the minimal amount of work that fakestore needs to do for a real snapd
client to be able to do it's registration tasks and setup an auto-refresh
request as it normally does during first boot of an Ubuntu Core system.

The hard coded values here are not the droids you are looking for.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>