tests/lib/fakestore: implement real store behavior with contextual refresh resp
This implements a more realistic behavior for the fakestore action endpoint,
with contextual refresh responses that are served depending on the client's
provided context. This lets us implement properly refreshes from a real snapd
client, as the fakestore can be setup with different revisions of the same snap
now, identified by specific revision numbers.

This also changes the response for the download-url's to use the request's Host
setting, as with this, we want to be able to use fakestore with a VM that will
not see the fakestore on localhost and instead will be pointed at 10.0.2.2 or
something similar, and so replying with localhost will be wrong, but the
request's Host setting will be correct.

Eventually we should eliminate the old hard-coded behavior for this endpoint,
but that requires changes to the test suite which are more than I intend to make
at this time.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>