overlord/snapstate: move locking to the backend
The two locks related to snap-confine and snap run (for refresh app
awareness), were acquired directly from the code in the snapstate
package. This logic has now moved to the backend.

The soft refresh check is calling a backend _function_ directly,
side-stepping fake backend, as that check is called from doInstall
which does not have access to the backend object yet.

I also fixed a bug in what is now called RunInhibitSnapForUnlink which
failed to effectively release the lock on failure, due to somewhat
tricky semantics of named return values in go.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>