boot/fdehook: add skeleton fdehook support
This commit adds minimal fdehook support with support for:
- InitialProvision (fde --initial-provision)
- Unlock (fde --unlock)
with systemd based confinement. Input is passed via json on stdin
and the sealed key data is read via stdout. The actual storing
of the sealed keys will be done by snapd and secboot. The hooks
only gets key input and is expected to generate key output.

In future versions there will be complex boot-chain data passed
to the hooks but this initial version is intentionally kept simple.
It will still be useful for the simple ARM fde based use-cases.

This work is based on the spike work from Claudio (thanks you!)
from
https://github.com/snapcore/snapd/compare/master...cmatsuoka:spike/fde-helper-tpm#diff-9d89d75d6aa1bcb1db124c226af0d8e0R43