tests: add tests for fsck
File system checker is an essential part of reliability of ubuntu-core
systems. The pair of new tests exercise fsck.vfat in particular and how
it is automatically invoked by the boot base, in a manner that is not
explored, automatically, in case a file system was not unmounted
cleanly.

Incidentally the test uncovered a real failure on core20, whatever
part is responsible for verifying vfat during the boot process, before
that partition is mounted, is not executing the same way it does on
core16 and core18.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>