tests/regression/lp-1898038/task.yaml: Fix test to pass on UC16/18
On Ubuntu Core 16/18 /etc/apparmor.d/cache is mounted from the hostfs so we
need to be more specific when checking for /etc/apparmor when looking at
the output from `mount` within the snap's mount namespace.

Also add an additional check to the test to ensure that the snaps have
correctly loaded their apparmor profiles to ensure apparmor_parser is
working correctly.