cmd/snap-confine: mask host's apparmor config
Some applications have access to the host's apparmor stack. On a system
with apparmor 3.x, a snap using even core20 as base, will have apparmor
2.x userspace tools available. If such application attempts to compile
an apparmor profile, then by the virtue of shared /etc, the host's
apparmor 3.x configuration will take effect.

Unfortunately, apparmor 2.x cannot parse apparmor 3.x configuration and
such applications fail to operate correctly. This notably affects Docker
and Multipass.

As a way out mask /etc/apparmor and /etc/apparmor.d from the shared host
/etc with those from the base snap.

For the record we tried to use snap-update-ns for this, but, as we
learned, we cannot do that using snap-update-ns. The mount profile
cannot know the revision of the base snap, because at the time the mount
profile is computed, the base snap may be unlinked. Subsequently
snap-update-ns cannot fill in this gap, as it may also run at the time
the base snap is unlinked (as in, has no current symlink). In contrast,
snap-confine runs only when the base snap is linked, so it can obtain
this information for free.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>