cmd/snap-bootstrap/initramfs-mounts: fsck ubuntu-seed on run mode boot
The ubuntu-seed partition is a vfat partition and is currently mounted writable,
which means that an unsafe shutdown/power-off will result it not being cleanly
unmounted which could lead to corruption. As such, we should always fun fsck on
it when booting to ensure that any dirtiness is cleaned up automatically on
boot.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>