cmd/snap-confine: mask host's apparmor config
* cmd/snap-confine: mask host's apparmor config

Some applications have access to the host's apparmor stack. On a system
with apparmor 3.x, a snap using even core20 as base, will have apparmor
2.x userspace tools available. If such application attempts to compile
an apparmor profile, then by the virtue of shared /etc, the host's
apparmor 3.x configuration will take effect.

Unfortunately, apparmor 2.x cannot parse apparmor 3.x configuration and
such applications fail to operate correctly. This notably affects Docker
and Multipass.

As a way out mask /etc/apparmor and /etc/apparmor.d from the shared host
/etc with those from the base snap.

For the record we tried to use snap-update-ns for this, but, as we
learned, we cannot do that using snap-update-ns. The mount profile
cannot know the revision of the base snap, because at the time the mount
profile is computed, the base snap may be unlinked. Subsequently
snap-update-ns cannot fill in this gap, as it may also run at the time
the base snap is unlinked (as in, has no current symlink). In contrast,
snap-confine runs only when the base snap is linked, so it can obtain
this information for free.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>

* tests/regression/lp-1898038: Add regression test for LP #1898038

This tests that /etc/apparmor bits get mounted inside the snap mount
namespace with either the docker-support and multipass-support interfaces
connected

* tests/regression/lp-1898038/task.yaml: Fix shellcheck errors

* tests/regression/lp-1898038/task.yaml: Fix test to pass on UC16/18

On Ubuntu Core 16/18 /etc/apparmor.d/cache is mounted from the hostfs so we
need to be more specific when checking for /etc/apparmor when looking at
the output from `mount` within the snap's mount namespace.

Also add an additional check to the test to ensure that the snaps have
correctly loaded their apparmor profiles to ensure apparmor_parser is
working correctly.

* tests/regression/lp-1898038: Rework for robustness and completeness

Ensure the test snaps can load an apparmor profile when the xxx-support
interface is connected and unload the profile at cleanup.

* tests/regression/lp-1898038: Rename test snaps as test-snapd-xxx

* tests/regression/lp-1898038: Use /bin/sh as shell interpreter

* tests/regression/lp-1898038: Use not instead of !

! will not actually stop execution of the test, we need to use `not`
instead

* tests/regression/lp-1898038: Keep shellcheck happy on the test scripts

* tests/regression/lp-1898038: Use not correctly in task.yaml

* tests/regression/lp-1898038: Test systems may not have aa-status so use it from the core snap

* tests/regression/lp-1898038: Run aa-status from within snap mount namespace

If we run it from the host mount namespace then it may be missing
dependencies etc from the default execution environment so instead enter
the snap's mount namespace and run it from there instead so it has
everything it may need.

* tests/regression/lp-1898038: Fix failing test on arch-linux-64

Run the snap and check for failure before trying to verify the state of the
snaps' mount namespace as this will create the mount namespace in the
process and also adds an extra check that the test behaves as expected.

* tests/regression/lp-1898038: Don't try and run on ubuntu-14.04

There is no nsenter for trusty which is required by this test so don't try
and run there otherwise we will fail.

* tests: tweak newlines for easier readability in lp-1898038

* snap-confine: add comment why /etc/apparmor{.d} gets mounted from the base instead of from the host

* cmd/snap-confine: (squash) fix apparmor profile

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>

* tests: adjust mount-ns after apparmor changes

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>

* cmd/snap-confine: fix spelling "minimum", indent

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>

* cmd/snap-confine: explain slave rules better

Previously they were joined together with normal/legacy mode rules and
it seemed like they belong to legacy rules. The reality is that they are
shared by both.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>

* tests: only compile, do not load apparmor profiles

This should allow this test to pass on systems without apparmor, and, in
fact, is exactly what the bug is about.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>

* tests: fake /etc/apparmor{,.d} for a specific regression test

This helps in running the test to the extent where it makes sense, but
it is not a very realistic behavior.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>

* tests: adjust test not to measure apparmor profiles

Profiles can only be loaded on some systems. We want to measure only
that the profiles can be compiled. In addition, since this is no longer
connection-oriented, simplify the test to ignore the disconnected
interface case.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>

* tests: exclude apparmor.d test from core16

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>

Co-authored-by: Alex Murray <alex.murray@canonical.com>
Co-authored-by: Michael Vogt <mvo@ubuntu.com>