osutil/disks: re-implement partition searching for disk w/ non-adjacent parts
We now have a bit more clarity on why this happens, and sometimes the kernel
either races with itself/udev in creating device nodes for the partitions, and
in case of clashing will create the clashing nodes with a blockext block device
instead, which will be a different major number than the normal mmc block
number.

Due to this, we now know we have to handle for sure the case where device
numbers are not-adjacent, so re-implement FindMatchingPartitionUUID as such. The
new implementation relies on partitions being subdirectories of the device path
in sysfs, and checks them for the partition file being "1". If it is "1", then
we know it belongs to the disk and is a partition so we can add it.

Also adjust the unit tests for this new implementation and switch to expecting
certain udevadm calls in order, as the order in which we query things matters
and the test becomes a bit more clear if we detail what is being queried in what
order.

Finally remove the test helper for mocking /sys/dev/block and instead use
dirs.SysfsDir which will automatically get updated whenever we change
dirs.GlobalRootDir as we do for each test.

Fixes: https://bugs.launchpad.net/snapd/+bug/1900842

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>