daemon: remove maintenance.json on startup, refactor internal var names
It is simpler and quicker to just delete the maintenance.json when we startup,
this saves time for the snap client reading/importing a file, as it can just
check if it doesn't exist and then move on.

restartSystem was overloaded to mean both a full system reboot and a simple
snapd restart. It's simpler to rename the variable on daemon to
activeRestartState, and then to use two variable in Stop() to track whether we
expect a restart at all (including a full system restart), and whether that
restart is a full system reboot or just a restart of snapd the daemon.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>