daemon/response.go: panic when marshalling a maintenance for state.RestartUnset
We now should delete/remove the maintenance.json file when it is unset, so we
can panic if we ever call maintenanceForRestartType on RestartUnset. This does
require adjusting ServeHTTP, to not call this if the Restarting() from state is
RestartUnset, which was the only current usage of this function with
RestartUnset now that we delete the maintenance.json file.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>