RFC: bootloader ping/pong protocol
This early patch is just to share the code and trigger a discussion.

We've recently came across a flaw in some versions of u-boot, where it
was unable to write to certain SD cards. During snapd kernel and boot
base update, snapd will instruct the bootloader to "try" to use specific
revisions of such snaps and reboot.

If the bootloader is unable to write to the persistent storage and the
device is eventually rebooted by external means, snapd will notice the
this fact and perform the rollback correctly.

What snapd doesn't know, however, is that the failure was caused by the
bootloader code, not by a flaw in one of the snaps.

In order to provide this bit of information, so that later on, snapd can
react accordingly (e.g. by reporting appropriate failure to the store
and attributing it to the gadget snap), it needs to detect this
condition. To do this a new *optional* protocol is proposed:

If the bootloader provides an environment variable
"snap_bootloader_ping" with the value "supported", then on environment
update, snapd will set that variable to "ping". If on subsequent boot,
that variable is still set to "ping" snapd sets it to broken and
proceeds normally (most likely to rollback the seemingly failed update).

The patch is shared extremely early, lacks tests, lacks core20 support.
Please advise on the direction.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>