secboot: don't fail if the volume was unlocked using the recovery key
What should we do if the volume was unlocked using the recovery key and
we can't retrieve the auth key from the kernel keyring?

We could load the value we saved beforehand but it feels like defeating
the purpose of having the auth key passed through the kernel keyring. But
if we say that in the case of unlocking with the recovery key the user
should also explicitly reseal also assumes that the user has interaction
capabilities to do so (such as a terminal) which doesn't seem to be
always the case. So, in the case of the PCR unlock failing on a secure
system without a user interface, should the system heal itself
automatically? This may have security implications too.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>