interfaces/builtin: Add `tmpfs-mount` interface
Multiple software projects make use of tmpfs as part of their
operation. Some notable examples are Ceph and Kubernetes.

We already have mount/umount of tmpfs support added as part of
the `kubernetes-support` interface and with this PR I propose
we add a separate interface for it so that multiple snaps
can consume it.

In reality it is a copy of the `cifs-mount` interface so it may
call for creating a more generic interface where fstype is a
dynamic property of the interface, let's discuss on the PR.