osutil/disks/mockdisk: panic if same mountpoint shows up again with diff opts
We should panic if we are passed a mapping that has the same mountpoint with
different options show up twice, as this is an inconsistent world-view, as one
mountpoint can only have one option set on it. The reason we have this option as
the input and not as a property of the disk itself is so that we can verify the
input, as whether or not to expect a decrypted device is a thing we need to
test, but it does have the unfortunate side effect where you can create
situations like this.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>