interfaces/builtin: Add device-mapper and LVM interfaces
This patch adds three separate interfaces for use with snaps
that require access consuming already set up device-mapper
devices or controling the device-mapper and/or LVM subsystems
to set up device-mapper devices.

The `device-mapper-devices` interface is suitable for snaps that
want to consume already prepared device-mapper devices, including
LVM Logical Volumes. A snap that would use the `block-devices`
interface but needs device-mapper/LVM should use this one.

The reason for not adding these rules to the already existing
`block-devices` interface is that the existing block-devices
interface specifically grants access to raw block devices and
explicitly prohibits access to partitions and any other mode of
access to block devices. Logical Volumes may come from underlying
devices that fall in the above category so adding it would be
circumventing the very rules the block-devices interface have
established.

The `device-mapper-control` interface can either be used
individually to allow a snap control over device-mapper pseudo
devices such as `dm-zero`. Normally it would probably be used
in conjunction with the `block-devices` interface to map actual
block devices into higher level device-mapper constructs.

The `lvm-control` interface should be used in conjunction with the
`device-mapper-*` and `block-devices` interfaces and is suitable
for snaps that need to take a raw block device and create LVM PVs,
VGs and LVs.