asserts: implement "storage-safety" in uc20 model assertion
This commit implements the new "storage-safety" header for the
model assertion. This header controls the encryption handling
for the given model and grade. Valid values for the storage-safety
header are "optional" and "encrypted". Note that the the
combination of "model: secured" and "storage-safety: optional"
is an error.

In addition to the assertion change this will need work on the
devicemanager to honor the settings and there will be a new
option for the gadget.yaml:
```
storage-safety-default: encrypted|unencrypted
```
Based on this default the recovery-chooser will offer defaults
about installing the device encrypted or unencrypted.