exportmanger: refactor Manifest and related types
This patch replaces Manifest, exportEntry and the ExportedSymlink types
with a new set of structures that reduce redundant information and
avoid hard-coding absolute paths. In addition, the new form clearly
highlights which side is sharing (snap world or host world) and which
side is consuming.

The new structure is as follows:

Manifest explicitly encodes export set name and version, in addition to
snap instance name and revision. The manifest also indicates if the
source is the host (explicit, rare) or a snap (implicit, common). A
manifest holds a map of named ExportSets. Previously those were encoded
directly in the ExportedSymlink, duplicating information.

An ExportSet has a name and a flag indicating if the consuming side is
another snap or the host. For now we only consume on the snap side but
this distinction is important both technically and for explicit clarity
that a single set is not sharing some content to one place and more
content to another place. Those can be modelled as distinct export sets.

Finally the ExportedFile structure has a name and a source path. The
name is relative to the export set name, which in turn is rooted in the
export directory specific to the exported name and version. The source
path is relative to the source, either the host or a snap.

Many methods that were previous on the leaf objects now become private
functions that take intermediate objects in order to construct the final
path.

Some terminology used throughout the code is streamlined. We only refer
to snap names when those really are snap names. We use exported name and
version in all the places where those are ambiguous.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>