gadget/many: rework detection of install created partitions, improve err msgs
We now only consider a structure on disk to be created at install (and thus a
candidate for deleting/recreating during install mode) if it is one of the roles
that is typically created at installs and if the start offset of the structure
on disk matches that of one that is to be created in the gadget.

Also, adjust the error message returned by ensureLayoutCompatibility to be
more specific about why two like-named partitions did not match, rather than the
existing message which just said that a on-disk structure could not be found.

Finally, adjust numerous tests to account for the absence of the
CreatedDuringInstall flag from the OnDiskStructure, because we cannot know if a
partition was created during install given only the disk layout, we need
additional information from the gadget so this flag is not useful and may be
dangerous in fact if used inappropriately.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>