cmd/snap-bootstrap/initramfs-mounts: initial robust recover mode behavior
This is the first take at making recover mode more robust against errors, and
trying to do things anyways even if some things fail on us. This will be
important for robust devices that are still able to get to a recover mode when
things go very very wrong in run mode, and gives snaps that are meant to rescue
the device from recover mode a chance to run and do something about fixing the
device.

This commit is missing unit tests.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>