exportmanager: simplify export manfiest
The manifest no longer holds snap instance + revision in addition to the
exported namd and version. The special case for things pretending to be
snapd are now reduced to a single optional field:
ExportedForSnapdAsRevision, which when set implies the exported name is
"snapd" and the version is the value of the field.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>