tests: fix rare interaction of tests.session and specific tests
I saw this message a few times but always forgot to fix the problem:

	+ echo 'stderr contains some messages'
	stderr contains some messages
	+ cat stderr.log
	/home/gopath/src/github.com/snapcore/snapd/tests/bin/tests.session:
	line 368: 19637 Terminated              stdbuf -oL dbus-monitor --system
	--monitor "$monitor_expr" > "$tmp_dir/dbus-monitor.pipe" 2>
	"$tmp_dir/dbus-monitor.stderr"
	+ exit 1

This is caused by the kill we use to stop dbus-monitor which is used to
ensure that dbus-monitor does not leak if it doesn't exit normally. When
we manage to kill it, mainly due to timing, we may produce a diagnostic
message from the shell, that a background task has terminated.

This patch silences the invocation of kill.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>