secboot: refactor UnlockVolumeUsingSealedKeyIfEncrypted to check disk enc first
This ensures that robust callers can always rely perfectly on the information in
the second return value to be accurate with what was attempted to be
introspected. Previously there was one return statement that didn't have the
information about whether there is a decrypted device or not, now the only ones
where we don't know that is when we fail to check the existence of the device
itself.

Note that we could even go the extra mile here and wrap the error values from
FindMatchingPartitionUUID, so that callers can inspect the return value for
i.e. matching disks.FilesystemLabelNotFoundError too, but we don't currently
need that so we are probably okay without it for now.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>