cmd/snap-bootstrap/initramfs-mounts: implement degraded mode for uc20 recovery
This implements degraded mode for UC20 recover mode. We will now fallback to
using the fallback sealed keys to unlock the data and save partitions if they
are encrypted and the run mode key is not usable.

This also implements placing a degraded.json file on /run/mnt/host/ubuntu-data
with information about what went wrong or where some things are.

Some remaining details to be hammered out:
- constant names in the JSON file
- code organization, the current one is very "iffy"
- console-conf is not yet disabled when we don't have the host ubuntu-data
  because when it is disabled it's very hard to debug, but we need to just copy
  the complete file to the right place to do so

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>