interfaces/fwupd: enforce the confined fwupd to align Ubuntu Core ESP layout
In the previous work, since UC16, we relied on the classic way to use
the path under /EFI/ubuntu for FW update, and this made shimx64.efi
redundant when it was copied to /EFI/ubuntu as well, as /EFI/boot
already provided bootx64.efi which is shim itself. It should be
improved by aligning the ESP layout the gadget specified, and the
confined fwupd just needs to do the same things under /EFI/boot instead.
However, the original rules for /EFI/ubuntu are needed to be maintained
for backward compatibility.