cmd/snap-bootstrap/initramfs-mounts: refactor degraded impl to use state machine
This refactors the degraded implementation to be more readable and
understandable. We use a state machine which runs one state at a time and
returns the next step so that each state is implemented by a single, simple
function rather than many nested if statements.

This also adds tests around degraded mode and the various situations we care
about.

Still to do are unit tests for the specific state machine logic functions.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>