secboot: add LockTPMSealedKeys() to lock access to keys independently
With the run mode and previous implementation of recover mode using
UnlockVolumeUsingSealedKeyIfEncrypted, we would always know what the last call
was and thus when to pass lockKeysOnFinish = true, however with the new degraded
mode, we do not know which call will be the last one.

As such, we need to use a separate function to lock access to the sealed keys,
and call that from the recover mode in initramfs-mounts.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>