cmd/snap-bootstrap/initramfs-mounts: use LockTPMSealedKeys in recover mode
We need to ensure that access to the sealed keys is always locked out, even when
we don't know what the last call to UnlockVolumeUsingSealedKeyIfEncrypted is.
The new implementation with the state machine suffers from this fact - we don't
know without a lot of state tracking, which specific call will be the last, so
instead just always defer a call to lock this.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>