cmd/s-b/initramfs-mounts: pack more information about degraded mode into json
This commit adds much more rich information about the partitions and errors that
happened during degraded setup while entering recover mode.

The degraded.json is now a more rich map, with each partition having it's own
map of strings to strings indicated state about them, such what device the
partition comes from, where it is mounted, if it was mounted, if it was
unlocked, what key it was unlocked with, as well as even if it was found at all.

Hopefully this is more future proof than the previous implementation which was
much more flat and less rich with more overloaded state keys.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>