interfaces/udev: use distro specific path to snap-device-helper
The snap-device-helper is located under $libexecdir which is /usr/lib/snapd on
some distros, but /usr/libexec/snapd on others (Fedora, recent openSUSE
Tumbleweed).

The commands triggered by rules are executed by udevd which runs in the same
mount ns as pid 1, which is the rootfs of a specific distribution.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>