cmd/snap-bootstrap/initramfs-mounts: use right dir for safe defaults in degraded
The dir we want is the /run/mnt/data/... one, which is what console-conf in
userspace will look at and check for the complete file. The one we were using
was the host's actual ubuntu-data one, which also means that this would
incorrectly mark console-conf as complete in run mode even if it hadn't yet run,
and still didn't prevent running console-conf from running in degraded recover
mode.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>