tests/core/failover: rm TODO:UC20, this test is already in kernel-failover
The test for UC20 kernel snap is done in tests/nested/core20/kernel-failover.
For the core variant of the test, we don't need to test this on UC20 since UC20
doesn't use the core snap as it's base snap.

However, we don't yet have a test for base snap failover, so add a TODO about
that to the kernel-failover test.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>