tests/main/lxd/prep-snapd-in-lxd.sh: wait for archive DNS before apt things
See for example the following failure:

+ echo 'Install snapd in container'
Install snapd in container
+ lxd.lxc exec my-ubuntu -- /root/prep-snapd-in-lxd.sh
+ [ -e /var/lib/dpkg/info/snapd.postrm ]
+ sed -i s#echo "Final directory cleanup"#umount /snap || true# /var/lib/dpkg/info/snapd.postrm
+ apt autoremove --purge -y snapd ubuntu-core-launcher

WARNING: apt does not have a stable CLI interface. Use with caution in scripts.

Reading package lists...
E: Type 'ubuntu' is not known on line 50 in source list /etc/apt/sources.list
E: The list of sources could not be read.
-----

I have not been able to reproduce this if I use the loop below, but when
debugging after this happens in the container by the time we get to a shell, apt
works normally for whatever reason.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>