include/lk: add snappy_boot_v2.h and snappy_boot_common.h
These include files will be used by bootloader implementations which marshal/
unmarshal these structures from partitions to communicate bootenv from snapd
to the bootloader and vice-versa.

These structures could be included in the original V1 struct by taking advantage
of the extra padding in the struct, but this would complicate using the struct,
as bootloader implementations would need to be coded to validate that the right
keys were being used, etc. for one version or another.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>