bootloader/lkenv: support v2 environment structs, specify v1 from lk.go
This does not yet use v2 environment structs, but enables them for use later on.

V2 structs are used when creating the lkenv, and many functions switch on which
object they actually operate on from which version of the lkenv is used.

The unit testing is not complete, but will be completed in a followup commit.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>