bootloader/lkenv: major refactor of lkenv
Refactor the lkenv package to use a better abstraction of the boot image
partition matrices that are common between the kernel revision for v1 and v2
run and the recovery system label for v2 recovery.

Also rename many methods to better match the naming scheme and consistently use
"boot image partitions" or "boot image partition labels" to refer to the first
row in the matrix, with value generically being the second row, which can either
be a kernel snap revision or a recovery system label.

Also make the signature for v2 recovery struct a separate value to be able to
indicate which lkenv object is which if we ever need to import or read an
arbitrary file, we can know which one it is from the bits alone.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>