daemon: fix reboot system action to not require a system label
We should be able to trigger a reboot without specifying the system label, since
we can detect the current system label and use that for the specified mode.

This was not caught before because of how we test daemon command endpoints and
the fact that the router we use operates weirdly for paths like /v2/systems,
where we have separate commands for the version without any sub-path and the
version with a sub-path.

Also fix a typo in an error message.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>