tests/main/lxd/prep-snapd-in-lxd.sh: wait for valid apt files before doing apt ops (#9663)
See for example the following failure:
```
+ echo 'Install snapd in container'
Install snapd in container
+ lxd.lxc exec my-ubuntu -- /root/prep-snapd-in-lxd.sh
+ [ -e /var/lib/dpkg/info/snapd.postrm ]
+ sed -i s#echo "Final directory cleanup"#umount /snap || true# /var/lib/dpkg/info/snapd.postrm
+ apt autoremove --purge -y snapd ubuntu-core-launcher

WARNING: apt does not have a stable CLI interface. Use with caution in scripts.

Reading package lists...
E: Type 'ubuntu' is not known on line 50 in source list /etc/apt/sources.list
E: The list of sources could not be read.
-----
```

It turns out that cloud-init in the LXD images is what is rewriting the sources.list file, so ideally we would just wait for cloud-init to be done with `cloud-init status --wait`, however that is currently broken in nested LXD containers, which we also test/support here, so instead we use the python snippet from @mvo5 which just waits up to 1 minute for the apt source files to be valid and then proceeds anyways so if the files never sort themselves out we won't wait forever and we will see in the logs that the apt sources never got rewritten.

Also eliminate a duplicated `apt install` command in another spread test.